package/freescale-imx/imx-seco: add the imx8dxl b0 revsision ahab image
Update the selection of the ahab container images for the imx8dxl
asics. The initial support assumed these were imx8 but this is
actually the imx8x architecture as well. This was fixed when run on
the first tests against the imx8dxl SoCs.

Files from upstream:
    mx8dxla0-ahab-container.img
    mx8dxla1-ahab-container.img
    mx8dxlb0-ahab-container.img

The B0 revision should be considered the latest

Signed-off-by: Charles Hardin <ckhardin@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>