package/alsa-plugins: fix dependencies for the alsa plugin package
The alsa-utils will remove and create the /usr/share/alsa dir
as a whole and since the alsa plugins needs to add config into
this same directory it should come after those install steps.

Also, the bluez alsa needs the plugins installed to find them
during the configuration phases - so, interlock the plugins
when configured.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>