package/xz: bump to version 5.4.3
This commit changes the version branch from 5.2.x to 5.4.x. This old
stable 5.2.x branch is expected to be end-of-life. The package
site [1] mention:

"""
5.4.3 was released on 2023-05-04. A minor bug fix release 5.2.12 to
the old stable branch was made on 2023-05-04. This is probably the
last release in the 5.2.x series.
"""

For a change log since 5.2.10, see [2].

This commit removes the package patch since the new version includes
alternate workarounds. See comment in [3].

The COPYING licence file hash has changed. A note about
Doxygen-generated HTML was added in [4].

COPYING.GPLv3 license file hash has also changed, as the file was
updated (http links changed by https) in [5].

[1] https://tukaani.org/xz/
[2] https://git.tukaani.org/?p=xz.git;a=blob;f=NEWS;h=7f83c81f61e8e6aa81525e44c072c76205eeb14b;hb=238b4e5458b4bd2cadefb768b8ea7c6b70a191ac
[3] https://github.com/tukaani-project/xz/pull/32#issuecomment-1438396969
[4] https://github.com/tukaani-project/xz/commit/f68f4b27f62f53fdac570885a1f4f23367ce6599
[5] https://github.com/tukaani-project/xz/commit/5a7b930efa7f9849d8da8397e8e5d8638f92be40

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>