package/acpica: remove HARDWARE_NAME make variable override
The HARDWARE_NAME Makefile variable was removed few years ago in
upstream (by Thomas). See [1].

This commit was first included in tag R01_19_17 (version 20170119).
See the change log [2].

This patch removes occurences of this HARDWARE_NAME, since it's now
unneeded. There is no change in the behavior of this build recipe.

[1] https://github.com/acpica/acpica/commit/0ba20f5f86d41bbe442eaeb426d13d25181fa512
[2] https://github.com/acpica/acpica/blob/R01_19_17/documents/changes.txt#L26

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>