package/gdal: drop remnants of cmake workarounds
In 2b43579e9453 (package/gdal: switch to cmake build to fix libgeotiff
detection) a workaround was added to use the generated 'Makefile' rather
than the bundled-for-autotools GNUMakefile, which was supposedly removed
for the then upcoming 3.6 version.

In 4c17985880be (package/gdal: bump version to 3.6.2) the bump occured,
but the workaround was left untouched. However, in 3.6.2, there is
indeed no GNUMakefile anymore.

Drop the workaround now.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>